updated layout and fixed build issue
