+++ app/constants/constant.jsx
@@ -1,0 +1,4 @@
+export const Metadata = {
+  title: "YOUR STORE",
+  description: "Created by RPG",
+};
