+++ app/pages/cart/page.jsx
@@ -1,14 +1,9 @@
 "use client";
 import React from "react";
-import MainLayout from "../page";
 import CartPage from "../../components/cartPage/CartPage";
 
 function page() {
-  return (
-    <MainLayout>
-      <CartPage />
-    </MainLayout>
-  );
+  return <CartPage />;
 }
 
 export default page;
