--- app/pages/login/page.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const page = () => {
-  return <div>login Page</div>;
-};
-
-export default page;
