--- app/pages/pagenotfound/page.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const page = () => {
-  return <div>page not found</div>;
-};
-
-export default page;
